# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Science Compiler API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/ingestion-runs:
    get:
      operationId: listIngestionRuns
      tags: [admin]
      summary: List recent ingestion runs
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of ingestion runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IngestionRun"
  /admin/ingestion-runs/{id}/results:
    get:
      operationId: getIngestionRunResults
      tags: [admin]
      summary: Get papers and claims created during a specific ingestion run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Papers and claims for the run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionRunResults"
        "404":
          description: Run not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/ingestion/run:
    post:
      operationId: triggerIngestion
      tags: [admin]
      summary: Trigger an immediate ingestion run
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerIngestionBody"
      responses:
        "200":
          description: Ingestion started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerIngestionResponse"
        "409":
          description: Ingestion already running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/ingestion-configs:
    get:
      operationId: listIngestionConfigs
      tags: [admin]
      summary: List ingestion configurations
      responses:
        "200":
          description: List of ingestion configs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IngestionConfig"
    post:
      operationId: createIngestionConfig
      tags: [admin]
      summary: Create a new ingestion config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIngestionConfigBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionConfig"
  /admin/ingestion-configs/{id}:
    patch:
      operationId: updateIngestionConfig
      tags: [admin]
      summary: Update an ingestion config
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIngestionConfigBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionConfig"
    delete:
      operationId: deleteIngestionConfig
      tags: [admin]
      summary: Delete an ingestion config
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /admin/review-queue:
    get:
      operationId: listReviewQueue
      tags: [admin]
      summary: Pending claims awaiting human review (low confidence or flagged)
      parameters:
        - name: limit
          in: query
          schema: { type: ["integer", "null"] }
        - name: offset
          in: query
          schema: { type: ["integer", "null"] }
      responses:
        "200":
          description: Pending claims
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewQueueResponse"
  /admin/claims/{id}/review:
    post:
      operationId: reviewClaim
      tags: [admin]
      summary: Approve, reject, or edit a pending claim
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewClaimBody"
      responses:
        "200":
          description: Updated claim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewClaimResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    IngestionRun:
      type: object
      properties:
        id:
          type: integer
        topicId:
          type: ["integer", "null"]
        topicName:
          type: ["string", "null"]
        status:
          type: string
        triggeredBy:
          type: string
        papersFound:
          type: integer
        papersProcessed:
          type: integer
        papersDeduplicated:
          type: integer
        fullTextFetched:
          type: integer
        lowConfidenceClaims:
          type: integer
        claimsExtracted:
          type: integer
        errorsCount:
          type: integer
        errorDetails:
          type: ["string", "null"]
        perSourceCounts:
          type: ["object", "null"]
          additionalProperties: { type: integer }
        startedAt:
          type: string
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, status, triggeredBy, papersFound, papersProcessed, claimsExtracted, errorsCount, startedAt, createdAt]
    IngestionRunResultPaper:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        authors:
          type: string
        journal:
          type: string
        publicationYear:
          type: integer
        methodologyType:
          type: string
        evidenceQuality:
          type: string
        pmid:
          type: ["string", "null"]
        doi:
          type: ["string", "null"]
        claimsCount:
          type: integer
        createdAt:
          type: string
      required: [id, title, authors, journal, publicationYear, methodologyType, evidenceQuality, claimsCount, createdAt]
    IngestionRunResultClaim:
      type: object
      properties:
        id:
          type: integer
        paperId:
          type: integer
        paperTitle:
          type: string
        claimText:
          type: string
        direction:
          type: string
        evidenceQuality:
          type: string
        population:
          type: string
        createdAt:
          type: string
      required: [id, paperId, paperTitle, claimText, direction, evidenceQuality, population, createdAt]
    IngestionRunResults:
      type: object
      properties:
        run:
          $ref: "#/components/schemas/IngestionRun"
        papers:
          type: array
          items:
            $ref: "#/components/schemas/IngestionRunResultPaper"
        claims:
          type: array
          items:
            $ref: "#/components/schemas/IngestionRunResultClaim"
      required: [run, papers, claims]
    IngestionConfig:
      type: object
      properties:
        id:
          type: integer
        topicId:
          type: integer
        topicName:
          type: ["string", "null"]
        pubmedQuery:
          type: string
        maxPapersPerRun:
          type: integer
        enabled:
          type: integer
        llmModel:
          type: string
        sources:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, topicId, pubmedQuery, maxPapersPerRun, enabled, llmModel, sources, createdAt, updatedAt]
    CreateIngestionConfigBody:
      type: object
      properties:
        topicId:
          type: integer
        pubmedQuery:
          type: string
          minLength: 1
          maxLength: 200
        maxPapersPerRun:
          type: ["integer", "null"]
        enabled:
          type: ["integer", "null"]
        llmModel:
          type: ["string", "null"]
        sources:
          type: ["array", "null"]
          items:
            type: string
      required: [topicId, pubmedQuery]
    UpdateIngestionConfigBody:
      type: object
      properties:
        pubmedQuery:
          type: ["string", "null"]
          maxLength: 200
        maxPapersPerRun:
          type: ["integer", "null"]
        enabled:
          type: ["integer", "null"]
        llmModel:
          type: ["string", "null"]
        sources:
          type: ["array", "null"]
          items:
            type: string
    PendingClaim:
      type: object
      properties:
        id:
          type: integer
        topicId:
          type: integer
        topicName:
          type: ["string", "null"]
        paperId:
          type: integer
        paperTitle:
          type: ["string", "null"]
        claimText:
          type: string
        direction:
          type: string
        evidenceQuality:
          type: string
        confidence:
          type: number
        status:
          type: string
        flagCount:
          type: integer
        createdAt:
          type: string
      required: [id, topicId, paperId, claimText, direction, evidenceQuality, confidence, status, flagCount, createdAt]
    ReviewQueueResponse:
      type: object
      properties:
        claims:
          type: array
          items:
            $ref: "#/components/schemas/PendingClaim"
        total:
          type: integer
      required: [claims, total]
...
    ReviewClaimBody:
      type: object
      properties:
        decision:
          type: string
          enum: [approve, reject, edit]
        notes:
          type: ["string", "null"]
          maxLength: 500
        edited:
          type: ["object", "null"]
          additionalProperties: true
      required: [decision]
    ReviewClaimResponse:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        confidence:
          type: number
      required: [id, status, confidence]
    TriggerIngestionBody:
      type: object
      properties:
        topicId:
          type: ["integer", "null"]
    TriggerIngestionResponse:
      type: object
      properties:
        message:
          type: string
        topicId:
          type: ["integer", "null"]
      required: [message]
...