# lib/api-zod/src/generated/types/createIngestionConfigBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Science Compiler API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateIngestionConfigBody {
  topicId: number;
  /**
   * @minLength 1
   * @maxLength 200
   */
  pubmedQuery: string;
  /** @nullable */
  maxPapersPerRun?: number | null;
  /** @nullable */
  enabled?: number | null;
  /** @nullable */
  llmModel?: string | null;
  /** @nullable */
  sources?: string[] | null;
}