# lib/api-zod/src/generated/types/updateIngestionConfigBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Science Compiler API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateIngestionConfigBody {
  /**
   * @maxLength 200
   * @nullable
   */
  pubmedQuery?: string | null;
  /** @nullable */
  maxPapersPerRun?: number | null;
  /** @nullable */
  enabled?: number | null;
  /** @nullable */
  llmModel?: string | null;
  /** @nullable */
  sources?: string[] | null;
}